Adds user and sudo_user variables in order to detect kube/config
This commit adds environment variables to the Makefile in order to
explicit determine that the user that is running the file is indeed
root.

The reasoning behind is that atomicapp now detects what user 'sudo' is
running on.

If 'sudo atomic run foo' is ran, it detects the user. However, because
this script is being ran as root/sudo, the .kube/config file is located
at /root/.kube/config and thus SUDO_USER and USER env variables need to
be explicitly implied.

Within the providers/kubernetes.sh file, kubectl config commands are
added in order to generate the .kube/config file.